[feat] setting화면에서 신청 받기
화면 전환 테스트 필요
